refactor: extract shared wallet-list admin handler
Deduplicate near-identical team-wallets and excluded-wallets endpoints
into a shared walletListGet/walletListPost in adminWalletList.ts.
Both route files are now thin config+delegation wrappers.
